refactor (#12343)
- enable running unit-tests
- refactor `walk.rs` also to enable more refactoring in future.
